fix: output directory detection
